Annotate max_delay as int

+++ 0x01-python_async_function/0-basic_async_syntax.py
@@ -7,7 +7,7 @@ import asyncio
 import random
 
 
-async def wait_random(max_delay: float = 10) -> float:
+async def wait_random(max_delay: int = 10) -> float:
     """Takes in an integer, and waits for a random delay between
     0 and max_delay seconds and eventually returns how long the wait was
     """
